+++ jamhh/src/be/ac/ulg/montefiore/run/jahmm/OpdfInteger.java
@@ -40,7 +40,7 @@ public class OpdfInteger
         probabilities = new double[nbEntries];
 
         for (int i = 0; i < nbEntries; i++) {
-            probabilities[i] = 1. / nbEntries;
+            probabilities[i] = 1. / ((double) nbEntries);
         }
     }
